requirements.txt
# pip install -r requirements.txt

python-dotenv

openai
pyupbit
ta
selenium
webdriver_manager
Pillow
youtube-transcript-api 
mysql-connector-python
streamlit
plotly